+++ pages/index.js
@@ -1,25 +1,81 @@
-import AudioDropzone from "@/components/AudioDropzone";
+import AudioPlayer from "@/components/AudioPlayer";
+import YouTubeComponent from "@/components/YTPlayer";
+import { Box, TextField, Typography } from "@mui/material";
+import { useState } from "react";
 
 export default function Home() {
+  const [audioSrc, setAudioSrc] = useState(null);
+  const [videoUrl, setVideoUrl] = useState(null);
+
+  const onDrop = (event) => {
+    event.preventDefault();
+
+    if (event.dataTransfer.items) {
+      for (let i = 0; i < event.dataTransfer.items.length; i++) {
+        if (event.dataTransfer.items[i].kind === "file") {
+          const file = event.dataTransfer.items[i].getAsFile();
+          const objectURL = URL.createObjectURL(file);
+          setAudioSrc(objectURL);
+        }
+      }
+    }
+  };
+
+  const onDragOver = (event) => {
+    event.preventDefault();
+  };
+
   return (
     <>
-      <div
-        style={{
-          // center everything
-          display: "flex",
-          justifyContent: "center",
-          alignItems: "center",
-          // make it pretty
-          height: "100vh",
-          width: "100vw",
-          backgroundColor: "#000000",
-          color: "#ffffff",
-          fontSize: "1rem",
-          fontFamily: "sans-serif",
-        }}
-      >
-        <AudioDropzone />
-      </div>
+      <Box>
+        {!audioSrc && !videoUrl && (
+          <div
+            onDrop={onDrop}
+            onDragOver={onDragOver}
+            style={{
+              border: "2px dashed #aaa",
+              padding: "20px",
+              position: "relative",
+              height: "200px",
+              display: "flex",
+              flexDirection: "column",
+              alignItems: "center",
+              justifyContent: "center",
+              borderRadius: "10px",
+            }}
+          >
+            <Typography variant="h5">Drop your audio file here</Typography>
+          </div>
+        )}
+        {!audioSrc && (
+          <div
+            style={{
+              border: "2px dashed #aaa",
+              padding: "20px",
+              position: "relative",
+
+              display: "flex",
+              flexDirection: "column",
+              alignItems: "center",
+              justifyContent: "center",
+              borderRadius: "10px",
+            }}
+          >
+            <TextField
+              id="outlined-basic"
+              label="YouTube URL"
+              variant="outlined"
+              value={videoUrl}
+              onChange={(event) => {
+                setVideoUrl(event.target.value);
+              }}
+            />
+          </div>
+        )}
+
+        {videoUrl && <YouTubeComponent videoUrl={videoUrl} />}
+        {audioSrc && <AudioPlayer audioSrc={audioSrc} />}
+      </Box>
     </>
   );
 }
